update readme and conf

+++ db/migrate/20141121190218_create_notations.rb
@@ -3,7 +3,7 @@ class CreateNotations < ActiveRecord::Migration
     create_table :notations do |t|
       t.belongs_to :exam
       t.integer :student
-      t.double :mark
+      t.float :mark
       t.text :results
       t.text :appreciation
